Moved api-attrs.h to flang/Common.
